# [Merged by Bors] - perf(analysis/inner_product_space/basic): speedups