Speed up all 30 declaration with elaboration time > 10s (on my machine) mostly by squeezing simps, except
```
elaboration of abs_inner_div_norm_mul_norm_eq_one_iff took 13.1s
```
is not sped up because I couldn't find a good way.

---

```
elaboration of finsupp.sum_inner took 20s
elaboration of finsupp.inner_sum took 10.3s
elaboration of dfinsupp.sum_inner took 10.1s
elaboration of dfinsupp.inner_sum took 10.5s
elaboration of inner_self_re_to_K took 14s
elaboration of inner_self_eq_norm_sq_to_K took 12.4s
elaboration of inner_mul_inner_self_le took 25.9s
elaboration of orthonormal.inner_left_sum took 13.7s
elaboration of orthonormal.inner_finsupp_eq_sum_left took 17.3s
elaboration of orthonormal.inner_finsupp_eq_sum_right took 18.5s
elaboration of orthonormal.linear_independent took 16.7s
elaboration of orthonormal.inner_finsupp_eq_zero took 12.2s
elaboration of orthonormal.orthonormal_of_forall_eq_or_eq_neg took 13s
elaboration of norm_eq_sqrt_inner took 11s
elaboration of norm_sub_sq took 20.7s
elaboration of orthonormal.equiv_refl took 11.3s
elaboration of orthonormal.equiv_symm took 15.4s
elaboration of orthonormal.equiv_trans took 16.6s
elaboration of norm_sub_eq_norm_add took 18.9s
elaboration of inner_eq_norm_mul_iff took 17.2s
elaboration of is_R_or_C.inner_product_space took 18.5s
elaboration of orthogonal_family.inner_right_dfinsupp took 11s
elaboration of orthogonal_family.inner_right_fintype took 11.8s
elaboration of orthogonal_family.inner_sum took 28s
elaboration of orthogonal_family.norm_sum took 13s
elaboration of orthogonal_family.orthonormal_sigma_orthonormal took 12.8s
elaboration of orthogonal_family.norm_sq_diff_sum took 15.6s
elaboration of orthogonal_family.independent took 36.7s
elaboration of direct_sum.is_internal.collected_basis_orthonormal took 14.8s
elaboration of inner_product_space.is_R_or_C_to_real took 17.3s
```

<!-- The text above the `---` will become the commit message when your
PR is merged. Please leave a blank newline before the `---`, otherwise
GitHub will format the text above it as a title.

To indicate co-authors, include lines at the bottom of the commit message 
(that is, before the `---`) using the following format:

Co-authored-by: Author Name <author@email.com>

Any other comments you want to keep out of the PR commit should go
below the `---`, and placed outside this HTML comment, or else they
will be invisible to reviewers.

If this PR depends on other PRs, please list them below this comment,
using the following format:
- [ ] depends on: #abc [optional extra text]
- [ ] depends on: #xyz [optional extra text]
-->

[![Open in Gitpod](https://gitpod.io/button/open-in-gitpod.svg)](https://gitpod.io/from-referrer/)